// types/project.ts
export interface Project {
    id: number
    title: string
    category: string
    image: string
    description: string
    technologies: string[]
    liveLink: string
    githubLink: string
    fullDescription?: string
    challenges?: string
    completedDate?: string
    role?: string
    client?: string
  }
  
  